dom/tests/mochitest/mouselock/file_syntheticMouseEvent.html

<!DOCTYPE HTML>
<html>
  <!--
  https://bugzilla.mozilla.org/show_bug.cgi?id=633602
  -->
  <head>
    <title>Test for Bug 633602</title>
    <script type="application/javascript"
      src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript"
      src="/tests/SimpleTest/EventUtils.js"></script>
    <script type="application/javascript"
      src="mouselock_utils.js"></script>

    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"
      />
  </head>
  <body>
    <a target="_blank"
      href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602</a>
    <div id="content">
      <img onclick="mClick++;" onmousedown="down++;" onmouseup="up++;"
        onmouseout="mOut++;" id="lockArea" width="101" height="101"
        alt="Synethetic Mouse Event Test Area"
      />
    </div>
    <pre id="test">
      <script type="application/javascript">
            /** Test for Bug 633602 **/
            /** Test to see if Synthetic Mouse behave the same
            regardless of mouse lock state **/
            var mClick = 0;
            var down = 0;
            var up = 0;
            var mOut = 0;
            var pointer = navigator.pointer;
            var lockArea;

            function runtest() {
              function mouseEvents() {
                sendMouseEvent({
                  type: "click"
                }, lockArea);
                sendMouseEvent({
                  type: "mousedown"
                }, lockArea);
                sendMouseEvent({
                  type: "mouseup"
                }, lockArea);
                sendMouseEvent({
                  type: "mouseout"
                }, lockArea);
              }
              document.addEventListener("mozfullscreenchange", function() {
                if (document.mozFullScreen) {
                  mouseEvents();
                  is(mClick, 1,
                    "Unlocked click: Synehtic mouse event failed.");
                  is(down, 1,
                    "Unlocked mousedown: Synehtic mouse event failed.");
                  is(up, 1,
                    "Unlocked mouseup: Synehtic mouse event failed.");
                  is(mOut, 1,
                    "Unlocked mouseout: Synehtic mouse event failed.");
                  pointer.lock(lockArea, function() {
                    mouseEvents();
                    is(mClick, 2,
                      "Locked click: Synehtic mouse event failed.");
                    is(down, 2,
                      "Locked mousedown: Synehtic mouse event failed.");
                    is(up, 2,
                      "Locked mouseup: Synehtic mouse event failed.");
                    is(mOut, 2,
                      "Locked mouseout: Synehtic mouse event failed.");
                    SimpleTest.finish();
                    document.mozCancelFullScreen();
                  }, function() {
                    ok(false, "Mouse lock failed to fire");
                    SimpleTest.finish();
                    document.mozCancelFullScreen();
                  });
                }
              });
              SimpleTest.waitForExplicitFinish();
              lockArea = document.getElementById("lockArea");
              lockArea.mozRequestFullScreen();
            }
            document.addEventListener("DOMContentLoaded", runtest, false);
      </script>
    </pre>
  </body>
</html>
